Refactor trace writer enqueue helper
